fix(wiggle): use float instead of integer for timestamps

+++ src/status_models/cluster_wiggle.rs
@@ -19,11 +19,11 @@ pub struct ClusterStoragePrimaryWiggle {
     pub finished_round: u16,
     pub finished_wiggle: u16,
 
-    pub smoothed_round_seconds: i64,
-    pub smoothed_wiggle_seconds: i64,
+    pub smoothed_round_seconds: f64,
+    pub smoothed_wiggle_seconds: f64,
 
-    pub last_round_finish_timestamp: i64,
+    pub last_round_finish_timestamp: f64,
     pub last_round_start_timestamp: f64,
-    pub last_wiggle_finish_timestamp: i64,
+    pub last_wiggle_finish_timestamp: f64,
     pub last_wiggle_start_timestamp: f64,
 }
